ImportKeys: Use data binding for extra informations about the key and refactoring
