+++ OpenKeychain/src/main/res/layout/import_keys_list_item_extra.xml
@@ -1,0 +1,36 @@
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <data>
+        <import type="java.util.ArrayList" />
+
+        <variable name="keyIdHex" type="String" />
+        <variable name="keyUserIds" type="ArrayList" />
+        <variable name="query" type="String" />
+        <variable name="revokedOrExpired" type="boolean" />
+    </data>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:textAppearance="?android:attr/textAppearanceSmall"
+            app:keyId="@{keyIdHex}"
+            app:keyRevokedOrExpired="@{revokedOrExpired}" />
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="4dp"
+            android:orientation="vertical"
+            app:keyRevokedOrExpired="@{revokedOrExpired}"
+            app:keyUserIds="@{keyUserIds}"
+            app:query="@{query}" />
+
+    </LinearLayout>
+
+</layout>
